pass exception as callback
